+++ API_Arcadia/API_Arcadia/Models/VetVisit.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public string Food { get; set; } = string.Empty;
         public float FoodWeight { get; set; }
-        public byte IdWeightUnit { get; set; }
+        public int IdWeightUnit { get; set; }
         public DateTime VisitDate { get; set; }
         public string Observations { get; set; } = string.Empty;
         public int IdAnimal { get; set; }
